+++ src/components/challenge/ChallengeCard/ChallengeCardStyled.tsx
