+++ src/components/challenge/ChallengeCard/index.tsx
